refactor: move mock data to dedicated files
