+++ src/app/(app)/products/[productId]/page.tsx
@@ -7,7 +7,8 @@ import { useEffect, useState } from 'react';
 import { notFound, useParams } from 'next/navigation';
 import ProductContainer from '@/components/products/detail/ProductContainer';
 import type { SimpleProductDetail, ProductSupplyChainLink, StoredUserProduct, DigitalProductPassport } from '@/types/dpp';
-import { USER_PRODUCTS_LOCAL_STORAGE_KEY, MOCK_DPPS } from '@/types/dpp';
+import { USER_PRODUCTS_LOCAL_STORAGE_KEY } from '@/types/dpp';
+import { MOCK_DPPS } from '@/data';
 import { Loader2 } from 'lucide-react';
 import { useToast } from "@/hooks/use-toast";
 import { syncEprelData } from '@/ai/flows/sync-eprel-data-flow';
