+++ src/app/api/products/route.ts
@@ -1,7 +1,7 @@
 
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { MOCK_DPPS } from '@/types/dpp';
+import { MOCK_DPPS } from '@/data';
 import type { DigitalProductPassport, DashboardFiltersState } from '@/types/dpp';
 
 export async function GET(request: NextRequest) {
