+++ src/app/api/v1/dpp/route.ts
@@ -4,7 +4,7 @@
 
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { MOCK_DPPS } from '@/types/dpp';
+import { MOCK_DPPS } from '@/data';
 import type { DigitalProductPassport, CustomAttribute, DashboardFiltersState } from '@/types/dpp';
 
 interface CreateDppRequestBody {
