+++ src/data/index.ts
@@ -1,0 +1,3 @@
+export * from './mockDpps';
+export * from './simpleMockProducts';
+export * from './mockSuppliers';
