+++ src/types/dpp/Product.ts
@@ -316,239 +316,6 @@ export interface DisplayableProduct {
 }
 
 // --- Mock data ---
-export const MOCK_DPPS: DigitalProductPassport[] = [
-  {
-    id: "DPP001",
-    productName: "EcoSmart Refrigerator X500",
-    category: "Appliances",
-    manufacturer: { name: "GreenTech Appliances"},
-    modelNumber: "X500-ECO",
-    metadata: { 
-      created_at: "2024-01-01T10:00:00Z", 
-      last_updated: "2024-07-30T10:00:00Z",
-      status: "published", 
-      dppStandardVersion: "CIRPASS v0.9 Draft" 
-    },
-    productDetails: {
-      description: "An eco friendly fridge.",
-      energyLabel: "A++",
-      imageUrl: "https://placehold.co/600x400.png",
-      imageHint: "refrigerator appliance",
-      materials: [{name: "Recycled Steel", percentage: 70, isRecycled: true}],
-      specifications: JSON.stringify({ "Capacity (Liters)": "400", "Annual Energy Consumption (kWh)": "150", "Noise Level (dB)": "38", "Dimensions (HxWxD cm)": "180x70x65", "Color": "Stainless Steel" }, null, 2),
-      customAttributes: [
-        {key: "Eco Rating", value: "Gold Star (Self-Assessed)"},
-        {key: "Special Feature", value: "AI Defrost Technology"},
-        {key: "Warranty Period", value: "5 Years"},
-        {key: "Country of Origin", value: "Germany"}
-      ],
-    },
-    compliance: {
-      eprel: { id: "EPREL_REG_12345", status: "Registered", url: "#eprel-link", lastChecked: "2024-01-18T00:00:00Z" }, 
-      esprConformity: { status: "conformant", assessmentId: "ESPR_ASSESS_001", assessmentDate: "2024-01-01" },
-      battery_regulation: { status: "not_applicable" },
-    },
-    ebsiVerification: { 
-      status: "verified", 
-      verificationId: "EBSI_TX_ABC123", 
-      lastChecked: "2024-07-25T00:00:00Z"
-    },
-    blockchainIdentifiers: { platform: "MockChain", anchorTransactionHash: "0x123abc456def789ghi012jkl345mno678pqr901stu234vwx567yz890abcdef"},
-    consumerScans: 1250,
-    lifecycleEvents: [
-      {id: "evt1", type: "Manufactured", timestamp: "2024-01-15T00:00:00Z", transactionHash: "0xabc...def", responsibleParty: "GreenTech Appliances"}
-    ],
-    certifications: [
-      {id: "cert1", name: "Energy Star", issuer: "EPA", issueDate: "2024-01-01T11:00:00Z", documentUrl: "#", transactionHash: "0xcertAnchor1", standard: "Energy Star Program Requirements for Refrigerators v6.0"},
-      {id: "cert2", name: "ISO 14001", issuer: "TUV Rheinland", issueDate: "2024-01-20T00:00:00Z", expiryDate: "2026-11-14", documentUrl: "#iso14001", vcId: "vc:iso:14001:greentech:dpp001", standard: "ISO 14001:2015"}
-    ],
-    documents: [
-      { name: "User Manual v1.2", url: "#manual_v1.2.pdf", type: "User Manual", addedTimestamp: "2024-01-15T00:00:00Z" },
-      { name: "Warranty Card", url: "#warranty.pdf", type: "Warranty", addedTimestamp: "2024-01-15T00:00:00Z" },
-    ],
-    supplyChainLinks: [
-      { supplierId: "SUP001", suppliedItem: "Compressor Unit XJ-500", notes: "Primary compressor supplier for EU market. Audited for ethical sourcing." },
-      { supplierId: "SUP002", suppliedItem: "Recycled Steel Panels (70%)", notes: "Certified post-consumer recycled content." }
-    ]
-  },
-  {
-    id: "DPP002",
-    productName: "Sustainable Cotton T-Shirt",
-    category: "Apparel",
-    manufacturer: { name: "EcoThreads"},
-    modelNumber: "ET-TS-ORG-M",
-    metadata: { last_updated: "2024-07-25T14:30:00Z", status: "draft", created_at: "2024-03-01T10:00:00Z" },
-    productDetails: {
-      description: "A sustainable t-shirt made from organic cotton.",
-      imageUrl: "https://placehold.co/600x400.png",
-      imageHint: "cotton t-shirt apparel",
-      materials: [{name: "Organic Cotton", percentage: 100}],
-      specifications: JSON.stringify({ "Fit": "Regular", "GSM": "180", "Origin": "India", "Care": "Machine wash cold" }, null, 2),
-      customAttributes: [{key: "Certifications", value: "GOTS, Fair Trade"}, {key: "Care Instructions", value: "Machine wash cold, tumble dry low"}]
-    },
-    compliance: {
-      eprel: { status: "Not Applicable", lastChecked: "2024-07-25T00:00:00Z" },
-      eu_espr: { status: "pending" },
-      battery_regulation: { status: "not_applicable" },
-    },
-    ebsiVerification: { status: "pending_verification", lastChecked: "2024-07-20T00:00:00Z"},
-    consumerScans: 300,
-    blockchainIdentifiers: { platform: "MockChain" },
-    certifications: [
-      {id: "cert3", name: "GOTS", issuer: "Control Union", issueDate: "2024-02-20", expiryDate: "2025-02-19", documentUrl: "#gots", standard: "Global Organic Textile Standard 6.0"},
-    ],
-    documents: [],
-    supplyChainLinks: [
-       { supplierId: "SUP003", suppliedItem: "Organic Cotton Yarn", notes: "GOTS Certified Supplier for all global production." }
-    ]
-  },
-  {
-    id: "DPP003",
-    productName: "Recycled Polymer Phone Case",
-    category: "Accessories",
-    manufacturer: { name: "ReCase It"},
-    modelNumber: "RC-POLY-IP15",
-    metadata: { last_updated: "2024-07-22T09:15:00Z", status: "flagged", created_at: "2024-04-10T10:00:00Z" },
-    compliance: {
-      eprel: { status: "Not Applicable", lastChecked: "2024-07-22T00:00:00Z" },
-      eu_espr: { status: "compliant" },
-      us_scope3: { status: "compliant" },
-      battery_regulation: { status: "not_applicable" },
-    },
-    consumerScans: 2100,
-     productDetails: { description: "A recycled phone case."},
-     blockchainIdentifiers: { platform: "OtherChain", anchorTransactionHash: "0x789polymerAnchorHash000333"},
-     documents: [],
-     supplyChainLinks: [],
-     ebsiVerification: { status: "not_verified", lastChecked: "2024-07-23T00:00:00Z"},
-  },
-  {
-    id: "DPP004",
-    productName: "Modular Sofa System",
-    category: "Furniture",
-    manufacturer: { name: "Comfy Living"},
-    modelNumber: "CL-MODSOFA-01",
-    metadata: { last_updated: "2024-07-20T11:00:00Z", status: "archived", created_at: "2023-12-01T10:00:00Z" },
-    compliance: {
-      eprel: { status: "Not Applicable", lastChecked: "2024-07-20T00:00:00Z" },
-      eu_espr: { status: "compliant" },
-      battery_regulation: { status: "not_applicable" },
-    },
-    consumerScans: 850,
-    productDetails: { description: "A modular sofa."},
-    documents: [],
-    supplyChainLinks: [],
-    ebsiVerification: { status: "error", lastChecked: "2024-07-19T00:00:00Z", message: "Connection timeout to EBSI node."},
-  },
-  {
-    id: "DPP005",
-    productName: "High-Performance EV Battery",
-    category: "Automotive Parts",
-    manufacturer: { name: "PowerVolt"},
-    modelNumber: "PV-EVB-75KWH",
-    metadata: { last_updated: "2024-07-29T08:00:00Z", status: "pending_review", created_at: "2024-05-01T10:00:00Z" },
-    productDetails: {
-      description: "A high-performance EV battery module.",
-      imageUrl: "https://placehold.co/600x400.png",
-      imageHint: "ev battery module",
-      specifications: JSON.stringify({ "Capacity (kWh)": "75", "Voltage (V)": "400", "Weight (kg)": "450", "Chemistry": "NMC 811" }, null, 2),
-      customAttributes: [
-        {key: "Cycle Life", value: "3000 cycles @ 80% DoD"},
-        {key: "Charging Time (0-80%)", value: "45 minutes (DC Fast Charge)"}
-      ]
-    },
-    compliance: {
-      eprel: { status: "Data Mismatch", lastChecked: "2024-07-28T00:00:00Z", id: "EPREL_OLD_ID" },
-      battery_regulation: {
-        status: "pending",
-        carbonFootprint: { value: 120, unit: "kg CO2e/kWh" },
-        recycledContent: [{ material: "Cobalt", percentage: 15 }],
-        stateOfHealth: {value: 98, unit: '%', measurementDate: "2024-07-15T00:00:00Z"},
-      },
-      eu_espr: { status: "pending" },
-    },
-    ebsiVerification: { status: "pending_verification", lastChecked: "2024-07-29T00:00:00Z"},
-    consumerScans: 50,
-    certifications: [
-      {id: "cert_bat_01", name: "UN 38.3 Transport Test", issuer: "TestCert Ltd.", issueDate: "2024-07-01", documentUrl: "#", transactionHash: "0xcertAnchorBat1", standard: "UN Manual of Tests and Criteria, Part III, subsection 38.3"}
-    ],
-    documents: [],
-    blockchainIdentifiers: { platform: "BatteryChain", anchorTransactionHash: "0xevBatteryAnchorHash555AAA"},
-    supplyChainLinks: []
-  },
-];
-export const SIMPLE_MOCK_PRODUCTS: SimpleProductDetail[] = [
-  {
-    id: "PROD001",
-    productName: "EcoFriendly Refrigerator X2000",
-    category: "Appliances",
-    status: "Active",
-    manufacturer: "GreenTech Appliances",
-    gtin: "01234567890123",
-    modelNumber: "X2000-ECO",
-    description: "State-of-the-art energy efficient refrigerator, built with sustainable materials and smart energy management. Features advanced frost-free systems and optimized airflow for even temperature distribution.",
-    imageUrl: "https://placehold.co/600x400.png",
-    imageHint: "modern refrigerator kitchen",
-    keySustainabilityPoints: ["Energy Star Certified", "Made with 70% recycled steel", "95% recyclable at end-of-life", "Low Global Warming Potential (GWP) refrigerant"],
-    keyCompliancePoints: ["EU Ecodesign Compliant", "EU Energy Labelling Compliant", "EPREL Registered"],
-    specifications: JSON.stringify({ "Capacity (Liters)": "400", "Annual Energy Consumption (kWh)": "150", "Noise Level (dB)": "38", "Dimensions (HxWxD cm)": "180x70x65", "Color": "Stainless Steel" }, null, 2),
-    customAttributes: [
-        {key: "Eco Rating", value: "Gold Star (Self-Assessed)"},
-        {key: "Special Feature", value: "AI Defrost Technology"},
-        {key: "Warranty Period", value: "5 Years"},
-        {key: "Country of Origin", value: "Germany"},
-    ],
-    materialsUsed: [
-      { name: "Recycled Steel", percentage: 70, source: "Certified Recycler A", isRecycled: true },
-      { name: "Bio-based Polymers (Insulation)", percentage: 15, source: "EcoPoly Corp" },
-      { name: "Tempered Glass (Shelves)", percentage: 10, source: "GlassWorld Inc." }
-    ],
-    energyLabelRating: "A+++",
-    repairability: { score: 8.5, scale: 10, detailsUrl: "#repair-details-PROD001" },
-    recyclabilityInfo: { percentage: 95, instructionsUrl: "#recycling-PROD001" },
-    supplyChainLinks: [
-      { supplierId: "SUP001", suppliedItem: "Compressor Unit XJ-500", notes: "Primary compressor supplier. Audited for fair labor practices." },
-      { supplierId: "SUP002", suppliedItem: "Recycled Steel Panels (70%)", notes: "Certified post-consumer recycled content." }
-    ],
-    complianceSummary: {
-      overallStatus: "Compliant",
-      eprel: { id: "EPREL12345", status: "Registered", url: "#eprel-link", lastChecked: "2024-07-01T00:00:00Z" },
-      ebsi: { status: "Verified", verificationId: "EBSI-VC-XYZ-00123", transactionUrl: "#ebsi-tx-001", lastChecked: "2024-07-05T00:00:00Z" },
-      specificRegulations: [
-        { regulationName: "EU Ecodesign 2019/2019", status: "Compliant", verificationId: "ECOD001", lastChecked: "2024-06-15T00:00:00Z", detailsUrl: "#ecodesign-report" },
-        { regulationName: "RoHS Directive 2011/65/EU", status: "Compliant", lastChecked: "2024-06-10T00:00:00Z" },
-        { regulationName: "WEEE Directive 2012/19/EU", status: "Compliant", lastChecked: "2024-06-10T00:00:00Z" },
-      ]
-    },
-    lifecycleEvents: [
-      { id: "lc001", eventName: "Manufacturing Complete", date: "2024-01-15T00:00:00Z", location: "EcoFactory, Germany", status: "Completed", iconName: "Factory" },
-      {
-        id: "lc002",
-        eventName: "Quality Assurance Passed",
-        date: "2024-01-16T00:00:00Z",
-        location: "EcoFactory, Germany",
-        status: "Completed",
-        iconName: "ShieldCheck",
-        keyDocuments: [
-          { name: "QA Report Q1-2024", type: "PDF", url: "#qa-report-q1-2024.pdf" },
-          { name: "Compliance Checklist (Internal)", type: "Link", url: "#compliance-checklist-internal" }
-        ]
-      },
-      { id: "lc003", eventName: "Shipped to Distributor", date: "2024-01-20T00:00:00Z", location: "Logistics Hub, Hamburg", status: "Completed", iconName: "Truck" },
-      { id: "lc004", eventName: "First Retail Sale", date: "2024-02-10T00:00:00Z", location: "Paris, France", status: "Completed", iconName: "ShoppingCart" },
-      { id: "lc005", eventName: "Scheduled Maintenance", date: "2025-02-15T00:00:00Z", notes: "Filter replacement due.", status: "Upcoming", iconName: "Wrench" },
-    ],
-    certifications: [
-      { name: "Energy Star", authority: "EPA", issueDate: "2024-01-01", documentUrl: "#", isVerified: true, standard: "Energy Star Program Requirements for Refrigerators v6.0", transactionHash: "0xcertAnchor1" },
-      { name: "ISO 14001", authority: "TUV Rheinland", issueDate: "2023-11-15", expiryDate: "2026-11-14", documentUrl: "#iso14001", isVerified: true, vcId: "vc:iso:14001:greentech:dpp001", standard: "ISO 14001:2015" }
-    ],
-    documents: [
-      { name: "User Manual v1.2", url: "#manual_v1.2.pdf", type: "User Manual", addedTimestamp: "2024-01-15T00:00:00Z" },
-      { name: "Warranty Card", url: "#warranty.pdf", type: "Warranty", addedTimestamp: "2024-01-15T00:00:00Z" },
-    ],
-  }
-];
-
 export const MOCK_PUBLIC_PASSPORTS: Record<string, PublicProductInfo> = {
   "PROD001": {
     passportId: "PROD001",
@@ -663,9 +430,3 @@ export const MOCK_PUBLIC_PASSPORTS: Record<string, PublicProductInfo> = {
   }
 };
 
-export const MOCK_SUPPLIERS: Supplier[] = [
-  { id: "SUP001", name: "GreenCompress Ltd.", contactPerson: "Sarah Miller", email: "sarah.miller@greencompress.com", location: "Stuttgart, Germany", materialsSupplied: "Eco-friendly compressors, Cooling units", status: "Active", lastUpdated: "2024-07-15T10:00:00Z" },
-  { id: "SUP002", name: "RecycleSteel Corp.", contactPerson: "John Davis", email: "jdavis@recyclesteel.com", location: "Rotterdam, Netherlands", materialsSupplied: "Recycled steel panels, Stainless steel components", status: "Active", lastUpdated: "2024-06-20T14:30:00Z" },
-  { id: "SUP003", name: "Organic Textiles Co.", contactPerson: "Aisha Khan", email: "akhan@organictextiles.com", location: "Coimbatore, India", materialsSupplied: "GOTS certified organic cotton yarn, Natural dyes", status: "Active", lastUpdated: "2024-07-01T09:00:00Z" },
-  { id: "SUP004", name: "PolySolutions Inc.", contactPerson: "Mike Chen", email: "chen.m@polysolutions.com", location: "Shanghai, China", materialsSupplied: "Recycled PET pellets, Bio-polymers, LED Chips & Drivers", status: "Pending Review", lastUpdated: "2024-05-10T11:00:00Z" },
-];
